+++ R/load_save.R
@@ -7,7 +7,7 @@
                        masterfile = NULL,
                        pkg_path = NULL) {
   .save_digest(new_data_digest, path = pkg_path)
-  flog.info("Saving to data")
+  .multilog_trace("Saving to data")
   # TODO get the names of each data object and save them separately. Provide a
   # function to load all.
   for (i in seq_along(object_names)) {