+++ R/processData.R
@@ -92,7 +92,7 @@ NULL
       winslash = "/"
     ), silent = TRUE)
   if (inherits(render_root, "try-error")) {
-    flog.error(paste0("render_root  = ", x, " doesn't exist."))
+    .multilog_error(paste0("render_root  = ", x, " doesn't exist."))
     # try creating, even if it's an old temp dir.
     # This isn't ideal. Would like to rather say it's a temporary
     # directory and use the current one..
@@ -129,7 +129,7 @@ DataPackageR <- function(arg = NULL, deps = TRUE) {
   # if it's an old temp dir, what then?
 
   if (!file.exists(target)) {
-    flog.fatal(paste0("Directory ", target, " doesn't exist."))
+    .multilog_fatal(paste0("Directory ", target, " doesn't exist."))
     {
       stop("exiting", call. = FALSE)
     }
@@ -144,13 +144,14 @@ DataPackageR <- function(arg = NULL, deps = TRUE) {
     dir.create(logpath, recursive = TRUE, showWarnings = FALSE)
     # open a log file
     LOGFILE <- file.path(logpath, "processing.log")
-    flog.appender(appender.tee(LOGFILE))
-    flog.info(paste0("Logging to ", LOGFILE))
+    .multilog_setup(LOGFILE)
+    .multilog_thresold(console = INFO, logfile = TRACE)
+    .multilog_trace(paste0("Logging to ", LOGFILE))
     # we know it's a proper package root, but we want to test if we have the
     # necessary subdirectories
     testme <- file.path(pkg_dir, c("R", "inst", "data", "data-raw"))
     if (!all(utils::file_test(testme, op = "-d"))) {
-      flog.fatal(paste0(
+      .multilog_fatal(paste0(
         "You need a valid package data strucutre.",
         " Missing ./R ./inst ./data or",
         "./data-raw subdirectories."
@@ -159,14 +160,14 @@ DataPackageR <- function(arg = NULL, deps = TRUE) {
         stop("exiting", call. = FALSE)
       }
     }
-    flog.info("Processing data")
+    .multilog_trace("Processing data")
     # read YAML
     ymlfile <- dir(
       path = pkg_dir, pattern = "^datapackager.yml$",
       full.names = TRUE
     )
     if (length(ymlfile) == 0) {
-      flog.fatal(paste0("Yaml configuration file not found at ", pkg_dir))
+      .multilog_fatal(paste0("Yaml configuration file not found at ", pkg_dir))
       {
         stop("exiting", call. = FALSE)
       }
@@ -174,19 +175,19 @@ DataPackageR <- function(arg = NULL, deps = TRUE) {
     ymlconf <- read_yaml(ymlfile)
     # test that the structure of the yaml file is correct!
     if (!"configuration" %in% names(ymlconf)) {
-      flog.fatal("YAML is missing 'configuration:' entry")
+      .multilog_fatal("YAML is missing 'configuration:' entry")
       {
         stop("exiting", call. = FALSE)
       }
     }
     if (!all(c("files", "objects") %in%
       purrr::map(ymlconf, names)[["configuration"]])) {
-      flog.fatal("YAML is missing files: and objects: entries")
+      .multilog_fatal("YAML is missing files: and objects: entries")
       {
         stop("exiting", call. = FALSE)
       }
     }
-    flog.info("Reading yaml configuration")
+    .multilog_trace("Reading yaml configuration")
     # files that have enable: TRUE
     assert_that("configuration" %in% names(ymlconf))
     assert_that("files" %in% names(ymlconf[["configuration"]]))
@@ -199,7 +200,7 @@ DataPackageR <- function(arg = NULL, deps = TRUE) {
       )
     ))
     if (length(r_files) == 0) {
-      flog.fatal("No files enabled for processing!")
+      .multilog_fatal("No files enabled for processing!")
       {
         stop("error", call. = FALSE)
       }
@@ -207,7 +208,7 @@ DataPackageR <- function(arg = NULL, deps = TRUE) {
     objects_to_keep <- purrr::map(ymlconf, "objects")[["configuration"]]
     render_root <- .get_render_root(ymlconf)
     if (!.validate_render_root(render_root)) {
-      flog.fatal(paste0(
+      .multilog_fatal(paste0(
         "Can't create, or render_root = ",
         render_root, " doesn't exist"
       ))
@@ -218,15 +219,14 @@ DataPackageR <- function(arg = NULL, deps = TRUE) {
 
     r_files <- file.path(raw_data_dir, r_files)
     if (all(!file.exists(r_files))) {
-      flog.fatal(paste0("Can't find any R or Rmd files."))
-      flog.fatal(paste0(
+      .multilog_fatal(paste0("Can't find any R or Rmd files."))
+      .multilog_fatal(paste0(
         "     Cant' find file: ",
         r_files[!file.exists(r_files)]
       ))
       stop("error", call. = FALSE)
     }
-    flog.info(paste0("Found ", r_files))
-    # TODO fix hidden warnings in test cases
+    .multilog_trace(paste0("Found ", r_files))
     old_data_digest <- .parse_data_digest(pkg_dir = pkg_dir)
     description_file <- normalizePath(file.path(pkg_dir, "DESCRIPTION"),
       winslash = "/"
@@ -241,7 +241,7 @@ DataPackageR <- function(arg = NULL, deps = TRUE) {
     # This is caught elsewhere
 
     if (length(objects_to_keep) == 0) {
-      flog.fatal("You must specify at least one data object.")
+      .multilog_fatal("You must specify at least one data object.")
       {
         stop("exiting", call. = FALSE)
       }
@@ -259,7 +259,7 @@ DataPackageR <- function(arg = NULL, deps = TRUE) {
       if (deps) {
         assign(x = "ENVS", value = ENVS, dataenv)
       }
-      flog.info(paste0(
+      .multilog_trace(paste0(
         "Processing ", i, " of ",
         length(r_files), ": ", r_files[i],
         "\n"
@@ -302,7 +302,7 @@ DataPackageR <- function(arg = NULL, deps = TRUE) {
       )
       # The created objects
       object_names <- ls(dataenv)
-      flog.info(paste0(
+      .multilog_trace(paste0(
         sum(objects_to_keep %in% object_names),
         " required data objects created by ",
         basename(r_files[i])
@@ -333,7 +333,7 @@ DataPackageR <- function(arg = NULL, deps = TRUE) {
       ) &
         string_check$isequal) {
         can_write <- TRUE
-        flog.info(paste0(
+        .multilog_trace(paste0(
           "Processed data sets match ",
           "existing data sets at version ",
           new_data_digest[["DataVersion"]]
@@ -353,7 +353,7 @@ DataPackageR <- function(arg = NULL, deps = TRUE) {
         pkg_description <- updated_version$pkg_description
         new_data_digest <- updated_version$new_data_digest
         can_write <- TRUE
-        flog.info(paste0(
+        .multilog_trace(paste0(
           "Data has been updated and DataVersion ",
           "string incremented automatically to ",
           new_data_digest[["DataVersion"]]
@@ -366,7 +366,7 @@ DataPackageR <- function(arg = NULL, deps = TRUE) {
         # edge case that shouldn't happen
         # but we test for it in the test suite
         can_write <- TRUE
-        flog.info(paste0(
+        .multilog_trace(paste0(
           "Data hasn't changed but the ",
           "DataVersion has been bumped."
         ))
@@ -376,7 +376,7 @@ DataPackageR <- function(arg = NULL, deps = TRUE) {
       )) {
         # edge case that shouldn't happen but
         # we test for it in the test suite.
-        flog.info(paste0(
+        .multilog_trace(paste0(
           "New DataVersion is less than ",
           "old but data are unchanged"
         ))
@@ -494,7 +494,7 @@ DataPackageR <- function(arg = NULL, deps = TRUE) {
         writeLines(text = save_docs[[i]], con = docfile)
       }
       close(docfile)
-      flog.info(
+      .multilog_trace(
         paste0(
           "Copied documentation to ",
           file.path(pkg_dir, "R", paste0(pkg_description$Package, ".R"))
@@ -509,7 +509,7 @@ DataPackageR <- function(arg = NULL, deps = TRUE) {
     # copy html files to vignettes
     .ppfiles_mkvignettes(dir = pkg_dir)
   }
-  flog.info("Done")
+  .multilog_trace("Done")
   return(can_write)
 }
 
@@ -743,7 +743,7 @@ document <- function(path = ".", install = TRUE) {
     to = file.path(path, "R", docfile),
     overwrite = TRUE
   )
-  flog.info("Rebuilding data package documentation.")
+  .multilog_trace("Rebuilding data package documentation.")
   devtools::document(pkg = path)
   location <- devtools::build(
     pkg = path, path = dirname(path),